conversion warning removed
git-svn-id: https://core.forge.funambol.org/svn/core/trunk/funambol/client-api/native@28844 e8e8ed6c-164c-0410-afcf-9e9a7c7d8c10

+++ test/common/spds/EncryptionTest.cpp
@@ -110,7 +110,7 @@ private:
 
         TransformationInfo infoe, infod;
 
-        infoe.size = strlen(clearText)*sizeof(char);
+        infoe.size = (long)strlen(clearText)*sizeof(char);
         infoe.password = password;
